разделил на модули
сделал выведение фоток одновременно
